feat: Add CategoriesController and cats-subcats-list view
This commit adds the `CategoriesController` class and the `cats-subcats-list.blade.php` view file. The `CategoriesController` is responsible for managing categories and subcategories, and the `cats-subcats-list.blade.php` view is used to display the list of categories and subcategories. These additions allow for better organization and management of categories and subcategories in the application.