feat: Add edit category functionality
The code changes include adding routes for editing and updating categories in the `admin.php` file, as well as adding the `edit-category.blade.php` view file. The `AdminCategoriesSubcategoriesList` component is also updated to include a link to the edit category page. This addition allows administrators to edit existing categories in the application.