feat: Add home page route
This commit adds the home page route to the web.php file. The route is defined as a GET request to the root URL and is handled by the `homePage` method in the `FrontEndController` class. This enhancement allows users to access the home page of the website, improving the overall user experience.