feat: Add route for showing purchase details
This commit adds a new route in the web.php file to handle the display of purchase details. The route is defined as a GET request to '/home/purchase/{id}' and is handled by the 'showPurchase' method in the 'FrontEndController' class. This feature allows users to view the details of a specific purchase, enhancing the overall user experience.